+++ proto-lens-benchmarks/benchmarks/UnusedFields.hs
@@ -21,7 +21,7 @@ import Control.DeepSeq (NFData)
 import Criterion.Main (Benchmark)
 import Lens.Family2 ((&), (.~))
 import Data.Int (Int32)
-import Data.ProtoLens.Message (def)
+import Data.ProtoLens.Message (defMessage)
 import Proto.UnusedFields
 import Proto.UnusedFields_Fields
 
@@ -38,10 +38,10 @@ defaultNumInt32s :: Int
 defaultNumInt32s = 10000
 
 populateFoo :: Int -> Int32 -> Foo
-populateFoo n k = def & field .~ replicate n k
+populateFoo n k = defMessage & field .~ replicate n k
 
 populateFooWithUnusedFields :: Int -> Int32 -> FooWithUnusedFields
-populateFooWithUnusedFields n k = def & field .~ replicate n k
+populateFooWithUnusedFields n k = defMessage & field .~ replicate n k
 
 benchmaker :: Int -> [Benchmark]
 benchmaker size =