+++ proto-lens-tests/tests/dependent_test.hs
@@ -1,6 +1,6 @@
 module Main (main) where
 
-import Data.ProtoLens (def)
+import Data.ProtoLens (defMessage)
 import Test.Framework (defaultMain)
 import Test.Framework.Providers.HUnit (testCase)
 import Test.HUnit ((@=?))
@@ -10,6 +10,6 @@ import Proto.Lib (LibMessage)
 
 main :: IO ()
 main = defaultMain
-    [ testCase "LibMessage" $ (def :: LibMessage) @=? def
-    , testCase "Dependent" $ (def :: Dependent) @=? def
+    [ testCase "LibMessage" $ (defMessage :: LibMessage) @=? defMessage
+    , testCase "Dependent" $ (defMessage :: Dependent) @=? defMessage
     ]