+++ proto-lens-tests/tests/package-deps_test.hs
@@ -1,6 +1,6 @@
 module Main where
 
-import Data.ProtoLens (def)
+import Data.ProtoLens (defMessage)
 import Lens.Family2 ((&), (.~), (^.))
 import Test.Framework.Providers.HUnit (testCase)
 import Test.HUnit ((@=?))
@@ -15,4 +15,4 @@ main = testMain [testWrapper]
 
 testWrapper :: Test
 testWrapper = testCase "testWrapper" $
-    42 @=? (def & foo . value .~ 42 :: Bar) ^. foo . value
+    42 @=? (defMessage & foo . value .~ 42 :: Bar) ^. foo . value