+++ proto-lens/src/Data/ProtoLens/Message.hs
@@ -30,7 +30,6 @@ module Data.ProtoLens.Message (
     FieldDefault(..),
     MessageEnum(..),
     -- * Building protocol buffers
-    Default(..),
     build,
     -- * Proto registries
     Registry,
@@ -51,7 +50,6 @@ module Data.ProtoLens.Message (
     ) where
 
 import qualified Data.ByteString as B
-import Data.Default.Class
 import Data.Int
 import qualified Data.Map as Map
 import Data.Map (Map)
@@ -71,12 +69,19 @@ import Data.ProtoLens.Encoding.Wire
 -- | Every protocol buffer is an instance of 'Message'.  This class enables
 -- serialization by providing reflection of all of the fields that may be used
 -- by this type.
-class Default msg => Message msg where
+class Message msg where
     -- | A unique identifier for this type, of the format
     -- @"packagename.messagename"@.
     messageName :: Proxy msg -> T.Text
+
+    -- | A message with all fields set to their default values.
+    --
+    -- Satisfies @encodeMessage defMessage == ""@ and @decodeMessage "" == Right defMessage@.
+    defMessage :: msg
+
     -- | The fields of the proto, indexed by their (integer) tag.
     fieldsByTag :: Map Tag (FieldDescriptor msg)
+
     -- | This map is keyed by the name of the field used for text format protos.
     -- This is just the field name for every field except for group fields,
     -- which use their Message type name in text protos instead of their
@@ -85,6 +90,7 @@ class Default msg => Message msg where
     fieldsByTextFormatName :: Map String (FieldDescriptor msg)
     fieldsByTextFormatName =
         Map.fromList [(n, f) | f@(FieldDescriptor n _ _) <- allFields]
+
     -- | Access the unknown fields of a Message.
     unknownFields :: Lens' msg FieldSet
 
@@ -142,8 +148,9 @@ data WireDefault value where
 
 -- | A proto3 field type with an implicit default value.
 --
--- This is distinct from 'Data.Default' to avoid orphan instances, and because
--- 'Bool' doesn't necessarily have a good Default instance for general usage.
+-- This is distinct from, say, 'Data.Default' to avoid orphan instances, and
+-- because 'Bool' doesn't necessarily have a good Default instance for general
+-- usage.
 class FieldDefault value where
     fieldDefault :: value
 
@@ -243,8 +250,8 @@ class (Enum a, Bounded a) => MessageEnum a where
 -- > x, y :: Lens' A Int
 -- > m :: A
 -- > m = build ((x .~ 5) . (y .~ 7))
-build :: Default a => (a -> a) -> a
-build = ($ def)
+build :: Message a => (a -> a) -> a
+build = ($ defMessage)
 
 -- | A helper lens for accessing optional fields.
 -- This is used as part of code generation, and should generally not be needed